Extract coins into own package
